requirements.txt
Django~=1.10.0
django-crispy-forms==1.6.1
django-ajax-selects

